requirements.txt
aiohttp==3.1.1
aiohttp-debugtoolbar==0.5.0
aiohttp-devtools==0.9
aiohttp-jinja2==0.17.0
aiohttp-security==0.2.0
aiohttp-session==2.3.0
aiopg==0.13.2
aioredis==1.1.0
alembic==0.9.9
async-timeout==2.0.1
attrs==17.4.0
chardet==3.0.4
click==6.7
hiredis==0.2.0
idna==2.6
idna-ssl==1.0.1
isort==4.3.4
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
multidict==4.1.0
passlib==1.7.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.2
python-editor==1.0.3
PyYAML==3.12
six==1.11.0
SQLAlchemy==1.2.6
uvloop==0.9.1
watchgod==0.1.1
yarl==1.1.1